Consolidate `AgentJanitor`
The `SimpleAgentJanitor` was only used as a base class. Consolidate into
one class `AgentJanitor` to simplify
